chor: put emddding function in one place
