+++ apps/backend/src/utils/helper.service.js
@@ -1,6 +1,12 @@
 import fs from "fs";
 import path from "path";
 import handlebars from "handlebars";
+import { GoogleGenerativeAI } from "@google/generative-ai";
+
+export const genAI = new GoogleGenerativeAI(process.env.GOOGLE_AI_API_KEY);
+
+// Model for converting text to vectors
+const embeddingModel = genAI.getGenerativeModel({ model: "embedding-001" });
 
 export const readTemplateFile = async (name, data) => {
     const content = fs
@@ -37,20 +43,63 @@ export const extractMetadata = (object) => {
     };
 };
 
-// Function to clean metadata before saving
-export function cleanMetadata (object) {
-    if (!object?.user) {
-        throw new Error("Invalid object: missing user reference");
-    }
+// Retry failed operations with exponential backoff
+export const withRetry = async (operation, maxRetries = 3, delay = 1000) => {
+    let lastError;
+    for (let attempt = 1; attempt <= maxRetries; attempt++) {
+        try {
+            return await operation();
+        } catch (error) {
+            lastError = error;
+            if (attempt === maxRetries) break;
 
-    return {
-        title: object.title?.trim() || "",
-        description: object.description?.trim() || "",
-        type: object.type?.toLowerCase() || "",
-        source: object.source?.toLowerCase() || "",
-        status: object.status?.toLowerCase() || "",
-        dueDate: object.dueDate ? new Date(object.dueDate).toISOString() : "",
-        isCompleted: Boolean(object.isCompleted),
-        userId: object.user.toString()
-    };
+            const shouldRetry = error.message.includes('503') ||
+                              error.message.includes('overloaded') ||
+                              error.message.includes('timeout');
+
+            if (!shouldRetry) throw error;
+
+            await new Promise(resolve => setTimeout(resolve, delay * Math.pow(2, attempt - 1)));
+        }
+    }
+    throw lastError;
+}
+
+// Convert text to vector representation for semantic search
+export const generateEmbedding = async (text) => {
+    return withRetry(async () => {
+        try {
+            const result = await embeddingModel.embedContent(text);
+            const embedding = result.embedding.values;
+
+            if (!embedding?.length) {
+                throw new Error("Invalid embedding generated");
+            }
+
+            return embedding;
+        } catch (error) {
+            console.error("Embedding generation failed:", error);
+            throw error;
+        }
+    });
+}
+
+export const generateEnhancedEmbedding = async (text, context = {}) => {
+    //  a rich contextual string by combining the text with relevant context
+    const contextualText = [
+        text,
+        context.type && `type:${context.type}`,
+        context.source && `platform:${context.source}`,
+        context.status && `status:${context.status}`,
+        context.dueDate && `due:${new Date(context.dueDate).toISOString()}`,
+        context.priority && `priority:${context.priority}`,
+        context.labels?.length && `labels:${context.labels.join(',')}`,
+        context.isCompleted && 'state:completed',
+        context.isArchived && 'state:archived',
+        context.relatedTasks?.map(task => `related:${task}`).join(' ')
+    ]
+        .filter(Boolean)
+        .join(' ');
+
+    return await generateEmbedding(contextualText);
 }
